requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
boto3>=1.34.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-magic>=0.4.27
aiofiles>=23.2.1
python-dotenv>=1.0.0
pypdf>=3.17.1
jinja2>=3.1.2
scikit-learn>=1.3.2
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.21.0
pandas>=2.0.0

# Development tools (optional)
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
ruff>=0.1.7
mypy>=1.7.1